+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog for the next release
 
+## Upgrade steps
+
 * Some unused thumbnail sizes have been deleted, run `./manage.py prune-thumbnails`
-  to delete dangling files 
+  to delete dangling files
+
+* The default location for ingredient images has changed. Please run 
+  `./manage migrate-ingredient-image-paths` to migrate existing entries. Note
+  that this is technically optional, as the old paths will continue working,
+  but it is advised for consistency.
+
+## Other changes
 
 * The ingredient API filter now supports range lookups on `nutriscore`
   (`gt`, `gte`, `lt`, `lte`) in addition to `exact` and `in`, enabling queries